Add prop handler to support existential bracket access on literals
The parser generates $ops: 'prop' directives to set properties on AST nodes
after creation. This is needed for existential bracket access on literal
arrays like [1,2,3]?[1], where the 'soak' property must be set to true on
the access node.

Without this handler, the code would generate invalid JavaScript with an
embedded error comment. With it, the code correctly generates:
  var ref;
  if ((ref = [1, 2, 3]) != null) {
    ref[1];
  }

This fixes 7 tests in test/old/12-existential.coffee.